refactor: initialization interface
